code/requirements.txt
google-cloud-bigquery==1.26.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==1.31.5
google-auth==1.35.0
google-auth-httplib2==0.0.4
google-cloud-core==1.7.2
google-cloud-logging==1.15.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
grpcio==1.44.0
idna==3.3
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2022.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
unidecode==1.3.8
pytest==8.0.0
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
scikit-learn==1.4.0
matplotlib==3.8.2
folium==0.15.1
openpyxl==3.1.2
python-dotenv==1.0.1